requirements.txt
# 核心依赖
opencv-python>=4.8.0
numpy>=1.24.0
pillow>=10.0.0

# OCR识别
pytesseract>=0.3.10
easyocr>=1.7.0

# 输入控制
pyautogui>=0.9.54
pynput>=1.7.6
pywin32>=306

# 机器学习（可选）
scikit-learn>=1.3.0
torch>=2.0.0
torchvision>=0.15.0

# 工具库
requests>=2.31.0
pyyaml>=6.0
pandas>=2.0.0
matplotlib>=3.7.0

# 日志和调试
loguru>=0.7.0
